Add Utilization popover with notification
